+++ src/ChefInterface.php
@@ -48,12 +48,12 @@ interface ChefInterface
     public function read(RecipeInterface $recipe): ChefInterface;
 
     /**
-     * To begin a sub recipe with the actual workplan.
+     * To reserve the current recipe to begin a sub recipe with the actual workplan.
      *
      * @param RecipeInterface $recipe
      * @return ChefInterface
      */
-    public function setAsideAndBegin(RecipeInterface $recipe): ChefInterface;
+    public function reserveAndBegin(RecipeInterface $recipe): ChefInterface;
 
     /**
      * To known when an ingredient missing in the work plan to start the cooking